+++ src/main/java/com/example/demo/repository/OrderDetailsRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.orm.OrderDetails;
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Long> {
     // Custom query methods can be added here if needed
-    List<OrderDetails> findByOrderId(int orderId);
+    List<OrderDetails> findByOrderId(Long orderId);
 
 }
